LibGfx: Sync to_skia_color_type
